Extract 'moduleSpecifiers' namespace out of importFixes (#24010)
